feat: 거래내역 필터링 기능을 위한 SearchFilter 컴포넌트 리팩토링
